feat: validation 애노테이션 추가

+++ todo/src/main/java/spring/todo/domain/Member.java
@@ -4,8 +4,11 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotEmpty;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
+import spring.todo.repository.member.MemberDto;
 
 @Entity
 @Getter @Setter
@@ -13,8 +16,7 @@ public class Member {
 
     @Id
     @GeneratedValue
-    private int id;
-
+    private Long id;
     @NotEmpty
     private String email;
     @NotEmpty
